requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0 